refactor: decouple infrastructure from application
