+++ src/Domain/Contracts/ClearContainerCacheInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace N3XT0R\XPub\Application\Cache;
+namespace N3XT0R\XPub\Domain\Contracts;
 
 
 interface ClearContainerCacheInterface
