+++ src/Infrastructure/Wordpress/Updater/PluginUpdateManager.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace N3XT0R\XPub\Infrastructure\Wordpress\Updater;
 
-use N3XT0R\XPub\Application\Cache\ClearContainerCacheInterface;
+use N3XT0R\XPub\Domain\Contracts\ClearContainerCacheInterface;
 use N3XT0R\XPub\Domain\Contracts\ReleaseProviderInterface;
 use N3XT0R\XPub\Shared\Plugin\PluginContext;
